[#2041] Move old munge tests to new_tests
